[librptexture] KhronosKTX2: Added a (uint32_t) cast to prevent a warning from ALIGN_BYTES().
The constness causes problems with __typeof__():

In file included from src/librptexture/stdafx.h:43,
                 from src/librptexture/fileformat/KhronosKTX2.cpp:14:
src/librptexture/fileformat/KhronosKTX2.cpp: In member function ‘void LibRpTexture::KhronosKTX2Private::loadKeyValueData()’:
src/librptexture/../librpbase/common.h:140:68: warning: type qualifiers ignored on cast result type [-Wignored-qualifiers]
  140 | # define ALIGN_BYTES(a, x) (((x)+((a)-1)) & ~((__typeof__(x))((a)-1)))
      |                                                                    ^
src/librptexture/fileformat/KhronosKTX2.cpp:616:8: note: in expansion of macro ‘ALIGN_BYTES’
  616 |   p += ALIGN_BYTES(4, sz);
      |        ^~~~~~~~~~~

+++ src/librptexture/fileformat/KhronosKTX2.cpp
@@ -613,7 +613,7 @@ void KhronosKTX2Private::loadKeyValueData(void)
 		}
 
 		// Next key/value pair.
-		p += ALIGN_BYTES(4, sz);
+		p += ALIGN_BYTES(4, (uint32_t)sz);
 	}
 }
